As Far As I've Managed
With much help from Steve and the internet, I've still only managed to pull off a loop that checks that an individual is not buying for themselves, and there are no repeats.  I haven't figured out a way to check vs SOs or do any sort of loop sampling from a dataframe (I have a project in process to try this with a dataframe, but it's been a huge pain in my ass).  Thoughts?  Advice?  Tutorial you know of out there?  I'm running out of time :(